FPAD-8430: Separate account deletion handler lambda and logic into dedicated files
