+++ src/handlers/account-deletion-processor-handler.ts
@@ -1,37 +1,9 @@
 import logger from '../commons/logger';
-import Ajv2019 from 'ajv/dist/2019';
-import { LOGS_PREFIX_SENSITIVE_INFO, MetricNames } from '../data-types/constants';
-import type { Context, SQSEvent, SQSRecord } from 'aws-lambda';
-import { addMetric, metric } from '../commons/metrics';
-import { accountDeleteMessageSchema } from '../contracts/account-delete-message';
-import { prettifyError } from 'zod';
-import jsonSafeParse from '../commons/json-safe-parse';
-import { AUTH_DELETE_ACCOUNTSchema } from '@govuk-one-login/event-catalogue-schemas';
-import { addEventMetadataToSchema } from '../commons/compile-schema';
-import { AccountStatusService, getPersistentAccountStatusService } from '../tables/account-status';
+import type { Context, SQSEvent } from 'aws-lambda';
+import { getPersistentAccountStatusService } from '../tables/account-status';
+import { processAccountDeletion } from './account-deletion-processor';
 
-const ajv2019 = new Ajv2019({ allErrors: true });
-
-const validateEventCatalogue = ajv2019.compile(
-  addEventMetadataToSchema(AUTH_DELETE_ACCOUNTSchema, 'AUTH_DELETE_ACCOUNT'),
-);
-
-enum ParserErrorType {
-  BODY_JSON_PARSER_ERROR = 'BODY_JSON_PARSER_ERROR',
-  BODY_FORMAT_PARSER_ERROR = 'BODY_FORMAT_PARSER_ERROR',
-  MISSING_USER_ID = 'MISSING_USER_ID',
-  EMPTY_USER_ID = 'EMPTY_USER_ID',
-}
-
-class ParserError extends Error {
-  constructor(
-    public readonly errorType: ParserErrorType,
-    message?: string,
-  ) {
-    super(message ?? 'The SQS message can not be parsed.');
-    this.name = 'ParserError';
-  }
-}
+const accountStatusService = getPersistentAccountStatusService();
 
 /**
  * Account deletion processor handler. Updates accounts and marks them for deletion.
@@ -42,86 +14,8 @@ class ParserError extends Error {
 export async function handler(
   event: SQSEvent,
   context: Context,
-  accountStatusService: AccountStatusService = getPersistentAccountStatusService(),
 ): Promise<void> {
   logger.addContext(context);
-  if (!event.Records[0]) {
-    logger.error('The event does not contain any records.');
-    return;
-  }
-
-  const updateRecordsByIdPromises = event.Records.map((record) => {
-    const userId = getUserId(record);
-    return userId ? updateDeleteStatusId(userId, accountStatusService) : undefined;
-  }).filter((prom) => prom !== undefined);
-  await Promise.all(updateRecordsByIdPromises);
-  metric.publishStoredMetrics();
+  await processAccountDeletion(event, accountStatusService)
 }
 
-/**
- * Function to parse the JSON message from the SQS Record.
- * @param record - The record passed in from the event.
- * @returns - User ID or undefined
- */
-function parseMessage(record: SQSRecord): string {
-  // JSON parse record.body
-  const recordBodyResult = jsonSafeParse(record.body);
-  if (!recordBodyResult.success) throw new ParserError(ParserErrorType.BODY_JSON_PARSER_ERROR);
-
-  // Validate against zod schema
-  const result = accountDeleteMessageSchema.safeParse(recordBodyResult.data);
-  if (!result.success)
-    throw new ParserError(
-      ParserErrorType.BODY_FORMAT_PARSER_ERROR,
-      `The SQS message can not be parsed. ${prettifyError(result.error)}`,
-    );
-
-  if (!validateEventCatalogue(recordBodyResult.data)) {
-    logger.info(`${LOGS_PREFIX_SENSITIVE_INFO} Event has failed event catalogue schema validation.`, {
-      validationErrors: validateEventCatalogue.errors,
-    });
-    addMetric(MetricNames.INVALID_EVENT_RECEIVED_EVENT_CATALOGUE, undefined, undefined, {
-      EVENT_NAME: result.data.event_name,
-    });
-  }
-
-  return result.data.user.user_id;
-}
-
-/**
- * Function to take the User ID from the SQS Record.
- * @param record - The record passed in from the event.
- * @returns - User ID as a string, with whitespace removed.
- */
-function getUserId(record: SQSRecord) {
-  try {
-    return parseMessage(record);
-  } catch (error) {
-    if (error instanceof ParserError) {
-      logger.error(error.message);
-      addMetric(MetricNames.DELETE_EVENT_PARSER_ERROR, undefined, undefined, {
-        ERROR: error.errorType,
-      });
-      metric.publishStoredMetrics();
-      return;
-    }
-
-    throw error;
-  }
-}
-
-/**
- * Function to call DynamoDB and mark the account as deleted.
- * @param userId - User ID taken from the record of the SQS Event.
- * @throws - Error if there is a problem updating the account.
- */
-async function updateDeleteStatusId(userId: string, accountStatusService: AccountStatusService) {
-  try {
-    await accountStatusService.updateDeleteStatus(userId);
-  } catch (error) {
-    logger.error(`${LOGS_PREFIX_SENSITIVE_INFO} Error updating account ${userId}`, { error });
-    addMetric(MetricNames.MARK_AS_DELETED_FAILED);
-    metric.publishStoredMetrics();
-    throw new Error('Failed to update the account status.', { cause: error });
-  }
-}
